move StateReader from state to kv
